+++ src/language-server/services/shared/__tests__/samples/sample.js
@@ -1,0 +1,11 @@
+/* eslint-disable @typescript-eslint/explicit-function-return-type */
+
+let mock = undefined;
+
+export function setMock(fn) {
+    mock = fn;
+}
+
+export function activate(context) {
+    mock.call(undefined, context);
+}